refactor: centralize atmosphere utilities
