+++ analysis/atmosphere/__init__.py
@@ -1,6 +1,7 @@
 """Atmosphere analysis utilities."""
 
 from .feature_extractor import AtmosphereFeatures
+from .market_air_sensor import MarketSnapshot, air_index
 from .regime_classifier import RegimeClassifier
 from .score_calculator import AtmosphereScore
 
@@ -8,4 +9,6 @@ __all__ = [
     "AtmosphereFeatures",
     "AtmosphereScore",
     "RegimeClassifier",
+    "MarketSnapshot",
+    "air_index",
 ]
