+++ piphawk_ai/vote_arch/__init__.py
@@ -1,10 +1,10 @@
 """Majority-vote trading pipeline components."""
+from analysis.atmosphere.market_air_sensor import MarketSnapshot, air_index
 from signals.mode_selector_v2 import select_mode
 
 from .ai_entry_plan import EntryPlan, generate_plan
 from .ai_strategy_selector import select_strategy
 from .entry_buffer import PlanBuffer
-from .market_air_sensor import MarketSnapshot, air_index
 from .pipeline import PipelineResult, run_cycle
 from .post_filters import final_filter
 from .regime_detector import MarketMetrics, rule_based_regime
