+++ piphawk_ai/vote_arch/market_air_sensor.py
@@ -1,21 +1,6 @@
-"""Calculate market air index used in prompts."""
-from __future__ import annotations
+"""Wrapper for market air utilities (deprecated)."""
+from analysis.atmosphere.market_air_sensor import MarketSnapshot, air_index
 
-from dataclasses import dataclass
-
-
-@dataclass
-class MarketSnapshot:
-    atr: float
-    news_score: float
-    oi_bias: float
-
-
-def air_index(m: MarketSnapshot) -> float:
-    """Return market air index based solely on technical indicators."""
-    vol_heat = min(m.atr / 0.1, 1)
-    flow_heat = abs(m.oi_bias)
-    # ニューススコアは一時的に無視する
-    return 0.7 * vol_heat + 0.3 * flow_heat
+# 互換性維持のための薄いラッパー
 
 __all__ = ["MarketSnapshot", "air_index"]
