+++ tests/test_pipeline_mode_integration.py
@@ -1,8 +1,8 @@
 import pytest
 
+from analysis.atmosphere.market_air_sensor import MarketSnapshot
 from piphawk_ai.vote_arch.ai_entry_plan import EntryPlan
 from piphawk_ai.vote_arch.entry_buffer import PlanBuffer
-from piphawk_ai.vote_arch.market_air_sensor import MarketSnapshot
 from piphawk_ai.vote_arch.pipeline import PipelineResult, run_cycle
 from piphawk_ai.vote_arch.regime_detector import MarketMetrics
